+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
-### 1.0.3 - 1.12.2016
-* Added support for android testing
+### 1.0.4 - 14.12.2016
+* Added support for XPath selectors
+* Added support for clicking elements
+* Added support for back button in android